changed z_margin behavior + refactor
the z_margin is only taken into account during a scan

moved a bunch of functions and keyboard.py to a util file
