+++ python/Sashimi/sashimi/helicon_stack.py
@@ -1,6 +1,5 @@
 import os
 import sys
-import time
 from pathlib import Path
 from glob import glob
 from PIL import Image
